Produto -> Locais de Armazenamento

+++ database/migrations/2017_12_11_090644_create_checklist_warehouse_quantities_table.php
@@ -14,12 +14,12 @@ class CreateChecklistWarehouseQuantitiesTable extends Migration
     public function up()
     {
         Schema::create('checklist_warehouse_quantities', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->unsignedInteger('warehouse_id');
+            $table->unsignedBigInteger('warehouse_id');
             $table->foreign('warehouse_id')->references('id')->on('warehouses')->onDelete('cascade');
 
-            $table->unsignedInteger('checklist_product_id');
+            $table->unsignedBigInteger('checklist_product_id');
             $table->foreign('checklist_product_id')->references('id')->on('checklist_products')->onDelete('cascade');
 
             $table->unsignedDecimal('quantity', 9, 3);
